NTUSER.DAT
This is the Official 1809  Default Hive From the send Official release of the 1809 Windows 10 Build, not the first one that was released and then pulled. This is for reference For comparison of your Default User hive to use a comparison. You can see the original create date. This was extracted from the Windows 10 ISO